requirements.txt
scipy
bezier
webcolors
matplotlib
seaborn
scikit-image
shapely
numpy
pandas
python-dotenv
gymnasium
tensorboard
wandb
stable-baselines3[extra]
torch
torchvision
tensorboard
wandb
# h5py
# datasets
# transformers
# jsonlines
# stable-baselines3
# torch==2.0.0+cu117
# torchvision==0.15.1+cu117